requirements.txt
# Libraries to be installed initially
# python -m pip install detectron2 -f \https://dl.fbaipublicfiles.com/detectron2/wheels/cu102/torch1.8/index.html
# sudo apt install build-essential libdbus-glib-1-dev libgirepository1.0-dev
# sudo apt-get install libudev-dev libsystemd-dev
# sudo apt-get install libcups2-dev
# sudo apt-get install libxml2-dev libxmlsec1-dev
# sudo apt install libcairo2-dev pkg-config python3-dev
# sudo apt-get -y install duplicity
# sudo apt-get install cmake
# pip install numpy==1.22.0
# pip install Cython

# conda config --add channels conda-forge
# conda config --set channel_priority strict

# Gstreamer packages
# conda install gstreamer
# conda install gst-plugins-bad
# conda install gst-plugins-good
# conda install gst-plugins-ugly
# conda install gst-plugins-base
# conda install gst-libav

absl-py==1.3.0
acme==1.1.0
altair==4.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
av==10.0.0
backcall==0.2.0
#backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
black==21.4b2
blinker==1.4
boto3==1.26.15
botocore==1.29.15
cachetools==5.2.0
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.0
cmake==3.25.0
colorama==0.4.3
commonmark==0.9.1
ConfigArgParse==0.13.0
configobj==5.0.6
contourpy==1.0.6
cryptography==2.8
cvxpy==1.2.2
cycler==0.11.0
#Cython
cython-bbox==0.1.3
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.4
Deprecated==1.2.13
# detectron2==0.6+cu102
distro==1.4.0
distro-info==1.0
dlib==19.24.0
docker-pycreds==0.4.0
easydict==1.10
ecos==2.0.10
entrypoints==0.3
exceptiongroup==1.0.4
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.8.0
filterpy==1.4.5
fire==0.4.0
flake8==5.0.4
flatbuffers==22.10.26
fonttools==4.38.0
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
gdown==4.5.3
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
html5lib==1.0.1
httplib2==0.14.0
hydra-core==1.2.0
idna==2.8
imageio==2.22.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
iopath==0.1.9
ipfsapi
ipython==8.6.0
isort==4.3.21
jedi==0.18.1
Jinja2==2.10.1
jmespath==1.0.1
joblib==1.2.0
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
lap==0.4.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lmdb==1.4.0
lockfile==0.12.2
louis==1.3
#lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mtcnn==0.1.1
mypy-extensions==0.4.3
nanoid==2.0.0
nats-py==2.2.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==2.8.8
#numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
olefile==0.46
omegaconf==2.2.3
onnx==1.12.0
onnx-simplifier==0.4.10
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
opt-einsum==3.3.0
osqp==0.6.2.post5
packaging==21.3
pandas==1.5.2
parameterized==0.8.1
paramiko==2.6.0
parsedatetime==2.4
parso==0.8.3
pathspec==0.10.2
pathtools==0.1.2
patsy==0.5.3
pbr==5.4.5
pexpect==4.6.0
pickleDB==0.9.2
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
portalocker==2.6.0
postgres==4.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
psycopg2-binary==2.9.5
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.5
pycairo==1.22.0
pycocotools==2.0.6
pycodestyle==2.9.1
pycparser==2.21
pycups==1.9.73
pydeck==0.8.0
pydot==1.4.2
pyflakes==2.5.0
Pygments==2.13.0
PyGObject==3.42.2
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyngrok==5.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytest==7.2.0
pytest-benchmark==4.0.0
python-dateutil==2.8.2
python-debian==0.1.36
python-dotenv==0.21.0
pytorchvideo==0.1.5
pytorchvideo-nightly==0.0.3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
qdldl==0.1.5.post2
redis==4.3.4
regex==2022.10.31
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
rich==12.6.0
rq==1.11.1
rsa==4.9
s3transfer==0.6.0
sahi==0.11.4
scikit-learn==1.1.1
scipy==1.7.3
scs==3.2.2
seaborn==0.12.1
SecretStorage==2.3.1
semver==2.13.0
sentry-sdk==1.11.0
setproctitle==1.3.2
Shapely==1.8.5.post1
shortuuid==1.0.11
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.6.1
statsmodels==0.13.5
streamlit==1.14.1
systemd-python==234
tabulate==0.9.0
tb-nightly==2.12.0a20221114
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.1
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminaltables==3.1.10
testresources==2.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
#torch==1.8.1
#torchaudio==0.8.1
#torchvision==0.9.1
torch
torchvision
torchaudio
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wadllib==1.3.3
wandb==0.13.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
yacs==0.1.8
yapf==0.32.0
zipp==3.10.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.5.2
